Extract ArgumentBinding into standalone class
