+++ core/trino-main/src/main/java/io/trino/sql/planner/optimizations/AddLocalExchanges.java
@@ -762,7 +762,7 @@ public class AddLocalExchanges
             // connector provided hash function
             verify(!(partitioningScheme.getPartitioning().getHandle().getConnectorHandle() instanceof SystemPartitioningHandle));
             verify(
-                    partitioningScheme.getPartitioning().getArguments().stream().noneMatch(Partitioning.ArgumentBinding::isConstant),
+                    partitioningScheme.getPartitioning().getArguments().stream().noneMatch(PartitioningArgument::isConstant),
                     "Table writer partitioning has constant arguments");
             PlanWithProperties newSource = source.accept(this, parentPreferences);
             PlanWithProperties exchange = deriveProperties(
@@ -798,7 +798,7 @@ public class AddLocalExchanges
             // connector provided hash function
             verify(!(partitioningScheme.getPartitioning().getHandle().getConnectorHandle() instanceof SystemPartitioningHandle));
             verify(
-                    partitioningScheme.getPartitioning().getArguments().stream().noneMatch(Partitioning.ArgumentBinding::isConstant),
+                    partitioningScheme.getPartitioning().getArguments().stream().noneMatch(PartitioningArgument::isConstant),
                     "Table writer partitioning has constant arguments");
 
             PlanWithProperties newSource = source.accept(this, defaultParallelism(session));
