+++ core/trino-main/src/main/java/io/trino/sql/planner/planprinter/NoOpAnonymizer.java
@@ -22,9 +22,9 @@ import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.type.Type;
 import io.trino.sql.planner.PartitioningHandle;
 import io.trino.sql.planner.Symbol;
+import io.trino.sql.planner.optimizations.PartitioningArgument;
 import io.trino.sql.tree.Expression;
 
-import static io.trino.sql.planner.Partitioning.ArgumentBinding;
 import static io.trino.sql.planner.plan.StatisticsWriterNode.WriteStatisticsTarget;
 import static io.trino.sql.planner.plan.TableWriterNode.WriterTarget;
 
@@ -68,7 +68,7 @@ public class NoOpAnonymizer
     }
 
     @Override
-    public String anonymize(ArgumentBinding argument)
+    public String anonymize(PartitioningArgument argument)
     {
         return argument.toString();
     }
